The Class class describes an Object class

Package: R.oo
Class Class

Object
~~|
~~+--Class

Directly known subclasses:

public static class Class
extends Object

The Class class describes an Object class. First of all, this class is most commonly used internally and neither the end user nor the programmer need to no about the class Class.

Class(name=NULL, constructor=NULL)

Arguments

name

Name of the class.

constructor

Constructor (function) of any Object class.

Fields and Methods

Methods:

$	-

$<-	-

.DollarNames	-

.subset2Internal	-

[[	-

[[<-	-

argsToString	Gets the arguments of a function as a character string.

as.character	Returns a short string describing the class.

forName	Gets a Class object by a name of a class.

getDetails	Lists the fields and methods of a class.

getFields	Returns the field names of a class.

getKnownSubclasses	Gets all subclasses that are currently loaded.

getMethods	Returns the method names of class and its super classes.

getName	Gets the name of the class.

getPackage	Gets the package to which the class belongs.

getRdDeclaration	Gets the class declaration in Rd format.

getRdHierarchy	Gets the class hierarchy in Rd format.

getRdMethods	Gets the methods of a class in Rd format.

getStaticInstance	Gets the static instance of this class.

getSuperclasses	Gets the super classes of this class.

isAbstract	Checks if a class is abstract or not.

isBeingCreated	Checks if a class is currently being initiated initiated.

isDeprecated	Checks if a class is deprecated or not.

isPrivate	Checks if a class is defined private or not.

isProtected	Checks if a class is defined protected or not.

isPublic	Checks if a class is defined public or not.

isStatic	Checks if a class is static or not.

newInstance	Creates a new instance of this class.

print	Prints detailed information about the class and its fields and methods.

Methods inherited from Object:
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, names, objectSize, print, save

Details

The class Class describes the Object class or one of its subclasses. All classes and constructors created by setConstructorS3() will be of class Class. Its methods provide ways of accessing static fields and static methods. Its print() method will print detailed information about the class and its fields and methods.

Author

Henrik Bengtsson