The Package class provides methods for accessing package information

Package: R.oo
Class Package

Object
~~|
~~+--Package

Directly known subclasses:

public class Package
extends Object

Creates a Package that can be thrown and caught. The Package class is the root class of all other Package classes.

Package(name=NULL)

Arguments

name

Name of the package.

Fields and Methods

Methods:

as.character	Gets a string representation of this package.

getAuthor	Gets the Author of this package.

getBundle	Gets the Bundle that this package might belong to.

getBundlePackages	Gets the names of the other packages that is in the same bundle as this package.

getChangeLog	Gets the change log of this package.

getClasses	Gets all classes of a package.

getContents	Gets the contents of this package.

getContribUrl	Gets the URL(s) from where this package can be installed.

getDataPath	Gets the path to the data (data/) directory of this package.

getDate	Gets the date when package was build.

getDescription	Gets the description of the package.

getDescriptionFile	Gets the description file of this package.

getDevelUrl	Gets the URL(s) from where the developers version of this package can be installed.

getDocPath	Gets the path to the accompanying documentation (doc/) directory of this package.

getEnvironment	Gets the environment of a loaded package.

getExamplePath	Gets the path to the example (R-ex/) directory of this package.

getHistory	-

getHowToCite	Gets the citation of this package.

getLicense	Gets the License of this package.

getMaintainer	Gets the Maintainer of this package.

getName	Gets the name of this package.

getNews	-

getPath	Gets the library (system) path to this package.

getPosition	Gets the search path position of the package.

getTitle	Gets the Title of this package.

getUrl	Gets the URL of this package.

getVersion	Gets the version of this package.

isLoaded	Checks if the package is installed on the search path or not.

isOlderThan	Checks if the package is older than a given version.

ll	Generates a list of informative properties of all members of the package.

load	Loads a package.

showChangeLog	Show the change log of this package.

showContents	Show the CONTENTS file of this package.

showDescriptionFile	Show the DESCRIPTION file of this package.

showHistory	-

showHowToCite	Show the HOWTOCITE file of this package.

showNews	-

startupMessage	Generates a 'package successfully loaded' package startup message.

unload	Unloads a package.

Methods inherited from Object:
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, names, objectSize, print, save

Author

Henrik Bengtsson

Examples

if (FALSE) # By defining .onAttach() as follows in zzz.R for a package, an
# instance of class Package with the same name as the package will
# be made available on the search path. More over, the code below
# will also inform the user that the package has been loaded:
#
#  > library(R.oo)
#  R.oo v0.52 (2003/04/13) was successfully loaded.
#
.onAttach <- function(libname, pkgname) {
  pkg <- Package(pkgname)
  assign(pkgname, pkg, pos=getPosition(pkg))
  cat(getName(pkg), " v", getVersion(pkg), " (", getDate(pkg), ")",
    " was successfully loaded.\n", sep="")
}

# The Package class works for any packages, loaded or not.

# Some information about the base package
pkg <- Package("base")
print(pkg)
#> [1] "Package: base v4.2.1 is loaded (pos=12).  Title: The R Base Package.  The official webpage is NA and the maintainer is R Core Team <do-use-Contact-address@r-project.org>.  The package is installed in /home/hb/shared/software/CBI/R-4.2.1-gcc9/lib/R/library/base/.  License: Part of R 4.2.1.  Description: Base R functions.  Type showNews(base) for package history, and ?base for help."
# [1] "Package: base v3.6.2 is loaded (pos=14).  Title: The R Base Package.
# The official webpage is NA and the maintainer is R Core Team <R-core@
# r-project.org>.  The package is installed in /usr/lib/R/library/base/.
# License: Part of R 3.6.2.  Description: Base R functions.  Type
# showNews(base) for package history, and ?base for help."
print(list.files(Package("base")$dataPath))
#> character(0)

# Some information about the R.oo package
print(R.oo::R.oo)
#> [1] "Package: R.oo v1.25.0-9002 is loaded (pos=2).  Title: R Object-Oriented Programming with or without References.  The official webpage is https://github.com/HenrikBengtsson/R.oo and the maintainer is Henrik Bengtsson.  The package is installed in /tmp/hb/Rtmpmrzm53/temp_libpath993724595dcf0/R.oo/.  License: LGPL (>= 2.1).  Description: Methods and classes for object-oriented programming in R with or without references.  Large effort has been made on making definition of methods as simple as possible with a minimum of maintenance for package developers.  The package has been developed since 2001 and is now considered very stable.  This is a cross-platform package implemented in pure R that defines standard S3 classes without any tricks.  Type showNews(R.oo) for package history, and ?R.oo for help."
# [1] "Package: R.oo v1.23.0-9000 .  Title: R Object-Oriented Programming
# with or without References.  The official webpage is https://github.com/
# HenrikBengtsson/R.oo and the maintainer is Henrik Bengtsson.  The package
# is installed in /home/alice/R/x86_64-pc-linux-gnu-library/3.6/R.oo/.
# License: LGPL (>= 2.1).  Description: Methods and classes for object-
# oriented programming in R with or without references.  Large effort has
# been made on making definition of methods as simple as possible with a
# minimum of maintenance for package developers.  The package has been
# developed since 2001 and is now considered very stable.  This is a
# cross-platform package implemented in pure R that defines standard S3
# classes without any tricks.  Type showNews(R.oo) for package history,
# and ?R.oo for help."