Class for converting Rdoc comments to Rd files

Package: R.oo
Class Rdoc

Object
~~|
~~+--Rdoc

Directly known subclasses:

public static class Rdoc
extends Object

Class for converting Rdoc comments to Rd files.

Rdoc()

Fields and Methods

Methods:

argsToString	Gets the arguments signature of a function.

check	Checks the compiled Rd files.

compile	Compile source code files containing Rdoc comments into Rd files.

createManPath	Creates the directory where the Rd files should be saved.

createName	Creates a class-method name.

declaration	Gets the class declaration.

escapeRdFilename	Escape non-valid characters in a filename.

getClassS4Usage	Gets the usage of a S4 class.

getKeywords	Gets the keywords defined in R with descriptions.

getManPath	Gets the path to the directory where the Rd files will be saved.

getNameFormat	Gets the current name format.

getObject	-

getPackageNameOf	Gets the package of a method or an object.

getRdTitle	Extracts the title string of a Rd file.

getUsage	Gets the usage of a method.

hierarchy	Gets the class hierarchy.

isKeyword	Checks if a word is a Rd keyword.

isVisible	Checks if a member is visible given its modifiers.

methodsInheritedFrom	Gets all methods inherited from a class in Rd format.

setManPath	Sets the path to the directory where the Rd files should be saved.

setNameFormat	Sets the current name format.

Methods inherited from Object:
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clearLookupCache, clone, detach, equals, extend, finalize, getEnvironment, getFieldModifier, getFieldModifiers, getFields, getInstantiationTime, getStaticInstance, hasField, hashCode, ll, load, names, objectSize, print, save

Author

Henrik Bengtsson

References

R developers, Guidelines for Rd files, https://developer.r-project.org/Rds.html, 2003

Examples