Rdoc$argsToString {R.oo}	R Documentation

Gets the arguments signature of a function

Description

Gets the arguments signature of a function.

Usage


## Static method (use this):
## Rdoc$argsToString(fcn, escapeRd=FALSE, collapse=TRUE, ...)

## Don't use the below:
## S3 method for class 'Rdoc'
argsToString(static, fcn, escapeRd=FALSE, collapse=TRUE, ...)

Arguments

fcn
A function.

escapeRd
If TRUE, certain Rd markup symbols are escaped.

collapse
If TRUE, each argument is returned as a single string, otherwise split up into a vector of strings as far as possible.

...
Not used.

Value

Returns a list of character strings.

Author(s)

Henrik Bengtsson

See Also

For more information see Rdoc.

[Package R.oo version 1.26.0 Index]