attach.BasicObject {R.oo}	R Documentation

Attach an BasicObject to the R search path

Description

Attach the members of an BasicObject to the R search path.

If trying to attach the same BasicObject twice without detaching it in between, a warning will be generated and nothing will be done.

Usage


## S3 method for class 'BasicObject'
attach(this, private=FALSE, pos=2, ...)

Arguments

private
If TRUE, private fields will also be attached, otherwise not.

pos
The position at in search path where the BasicObject should be inserted.

...
Not used.

Value

Returns TRUE if the BasicObject was attached, otherwise FALSE.

Author(s)

Henrik Bengtsson

See Also

*detach() and attach(), detach(). For more information see BasicObject.

[Package R.oo version 1.26.0 Index]