clone.Object {R.oo}	R Documentation

Clones an Object

Description

Creates an identical copy of the object and returns a reference to the new object.

Usage


## S3 method for class 'Object'
clone(this, ...)

Arguments

...
Not used.

Details

Please note that no constructors are called during the creation of the clone and neither is any static class code called.

Value

A reference to the new object.

Author(s)

Henrik Bengtsson

See Also

For more information see Object.

Examples


  o1 <- Object()
  o2 <- clone(o1)

  print(equals(o1, o2))

[Package R.oo version 1.26.0 Index]