Rdoc$compile {R.oo}	R Documentation

Compile source code files containing Rdoc comments into Rd files

Description

Compile source code files containing Rdoc comments into Rd files.

Note, the class and methods to be compiled have to be loaded into R by for instance library() or source() before calling this method.

Usage


## Static method (use this):
## Rdoc$compile(filename=".*[.]R$", destPath=getManPath(this), showDeprecated=FALSE,
##   addTimestamp=FALSE, locale="C", verbose=FALSE, source=FALSE, check=TRUE,
##   debug=FALSE, ...)

## Don't use the below:
## S3 method for class 'Rdoc'
compile(this, filename=".*[.]R$", destPath=getManPath(this), showDeprecated=FALSE,
  addTimestamp=FALSE, locale="C", verbose=FALSE, source=FALSE, check=TRUE, debug=FALSE,
  ...)

Arguments

filename
The pathname or filename pattern of the Rdoc files to be compiled.

destPath
The path where the generated Rd files should be saved.

showDeprecated
If TRUE, Rd files are generated for deprecated objects too, otherwise not.

addTimestamp
If TRUE, a date and time stamp is added to the Rd header comments. This timestamp might be confusing for version control systems, which is why it can be turned off with FALSE.

locale
The locale to be set/used when compiling Rdoc comments. This help assuring strings are sorted the same way across systems.

source
If TRUE, the Rdoc files will be source():ed first. This work of course only for Rdoc files that are R source files.

verbose
If TRUE, detailed compilation information is printed.

debug
If TRUE, extra debug information is printed.

...
Not used.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

For more information see Rdoc.

[Package R.oo version 1.26.0 Index]