compileRdoc {R.oo}	R Documentation

Compiles the Rdoc help comments in a package

Description

Compiles the Rdoc help comments in a package by parsing the source code files and writes Rd help files.

Usage


compileRdoc(pkgname=NULL, path=pkgname, ..., verbose=TRUE)

Arguments

pkgname
A character string specifying the package name. If NULL, the package name is inferred from the DESCRIPTION file that is automatically search for in the subdirectories.

path
A character string specifying the path of the package source directory.

...
Additional arguments passed to Rdoc\$compile().

verbose
If TRUE, verbose output is printed, otherwise not.

Details

To compile all Rdoc comments in a package 'PkgA' that is located under the current working directory (e.g. PkgA/R/*.R) from the system command line, do:


    Rscript -e R.oo::compileRdoc()
  

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

See the Rdoc class.

[Package R.oo version 1.26.0 Index]