Rdoc$createManPath {R.oo}	R Documentation

Creates the directory where the Rd files should be saved

Description

Creates the directory where the Rd files should be saved.

Usage


## Static method (use this):
## Rdoc$createManPath(...)

## Don't use the below:
## S3 method for class 'Rdoc'
createManPath(this, ...)

Arguments

...
Not used.

Value

Returns TRUE if the directory was creates, FALSE if it already exists and throws an Exception if failed.

Author(s)

Henrik Bengtsson

See Also

*getManPath() For more information see Rdoc.

[Package R.oo version 1.26.0 Index]