detach.Object {R.oo}	R Documentation

Detach an Object from the R search path

Description

Detach, from the R search path, an Object that has previously been attached. If the Object was not attached, a warning will be generated and nothing will be done.

Usage


## S3 method for class 'Object'
detach(this, ...)

Arguments

...
Not used.

Value

Returns TRUE if the Object was detached, otherwise FALSE.

Author(s)

Henrik Bengtsson

See Also

*attach() and attach(), detach(). For more information see Object.

Examples

## Not run: For a complete example see help(Object).

[Package R.oo version 1.26.0 Index]