equals.Object {R.oo}	R Documentation

Compares an object with another

Description

Compares an object with another and returns TRUE if they are equal. The equal property must be

1) reflexive, i.e. equals(o1,o1) should be TRUE.

2) symmetric, i.e. equals(o1,o2) is TRUE if and only if equals(o2,o1) is TRUE.

3) transitive, i.e. equals(o1,o2) is TRUE and equals(o2,o3) is TRUE, then equals(o1,o3) should be TRUE.

5) consistent, i.e. equals(o1,o2) should return the same result on multiple invocations as long as nothing has changed.

6) equals(o1,NULL) should return FALSE, unless o1 is also NULL.

By default, the method returns TRUE if and only if the two references compared refer to the same Object, i.e. ( !is.null(obj) && (hashCode(this) == hashCode(obj)) ).

Usage


## S3 method for class 'Object'
equals(this, other, ...)

Arguments

other
The other object this Object should be compared to.

...
Not used.

Value

Returns TRUE if the Object's are equal, otherwise FALSE.

Author(s)

Henrik Bengtsson

See Also

*hashCode(). For more information see Object.

Examples


  o1 <- Object()
  o2 <- clone(o1)
  equals(o1, o1)    # TRUE
  equals(o1, o2)    # FALSE

[Package R.oo version 1.26.0 Index]