extend.BasicObject {R.oo}	R Documentation

Extends another class

Description

via a mechanism known as "parasitic inheritance". Simply speaking this method "extends another class". What is actually happening is that it creates an instance of class name ...className, by taking another BasicObject instance and add ...className to the class list and also add all the named values in ... as fields to the new instance.

The method should be used by the constructor of a class and nowhere else.

Usage


## S3 method for class 'BasicObject'
extend(this, ...className, ...)

Arguments

...className
The name of new class.

...
Named values representing the fields of the new instance.

Value

Returns an BasicObject of class className.

Author(s)

Henrik Bengtsson

See Also

For more information see BasicObject.

[Package R.oo version 1.26.0 Index]