Class$forName {R.oo}	R Documentation

Gets a Class object by a name of a class

Description

Gets a Class object by a name of a class. If no such class exists and exception is thrown.

Usage


## Static method (use this):
## Class$forName(name, ...)

## Don't use the below:
## S3 method for class 'Class'
forName(static, name, ...)

Arguments

...
Optional arguments passed to internal lookup function.

Value

Returns a Class.

Author(s)

Henrik Bengtsson

See Also

class(). setConstructorS3(). For more information see Class.

Examples


  print(Class$forName("Exception"))

[Package R.oo version 1.26.0 Index]