getAuthor.Package {R.oo}	R Documentation

Gets the Author of this package

Description

Gets the Author of this package as specified by the DESCRIPTION file.

Usage


## S3 method for class 'Package'
getAuthor(this, as=c("character", "person"), include=c("given", "family"), ...)

Arguments

as
A character string specifying the return format.

include
A character vector specifying which person fields to include if returning a character string.

...
Optional arguments passed to format.person.

Value

Returns a character string or a person object.

Author(s)

Henrik Bengtsson

See Also

For more information see Package.

Examples


  pkg <- Package("R.oo")
  print(getAuthor(pkg))

[Package R.oo version 1.26.0 Index]