getBundle.Package {R.oo}	R Documentation

Gets the Bundle that this package might belong to

Description

Gets the Bundle that this package might belong to as specified by the DESCRIPTION file.

Usage


## S3 method for class 'Package'
getBundle(this, ...)

Arguments

...
Not used.

Details

The first call to this method is normally slow because all installed packages are scanned. The result of this first call is cached and used as the return value for all subsequent calls, which are then much faster.

Value

If the package is part of a bundle, the name of the bundle is returned. Otherwise, NULL is returned.

Author(s)

Henrik Bengtsson

See Also

For more information see Package.

Examples


  pkg <- Package("R.oo")
  print(getBundle(pkg))

[Package R.oo version 1.26.0 Index]