getBundlePackages.Package {R.oo}	R Documentation

Gets the names of the other packages that is in the same bundle as this package

Description

Gets the names of the other packages that is in the same bundle as this package as specified by the DESCRIPTION file.

Usage

## S3 method for class 'Package'
getBundlePackages(this, ...)

Arguments

...
Not used.

Details

The first call to this method is normally slow because all installed packages are scanned. The result of this first call is cached and used as the return value for all subsequent calls, which are then much faster.

Value

Returns a character vector of package names or NULL.

Author(s)

Henrik Bengtsson

See Also

For more information see Package.

Examples

  pkg <- Package("R.oo")
  print(getBundle(pkg))

[Package R.oo version 1.22.0 Index]