getDate.Package {R.oo}	R Documentation

Gets the date when package was build

Description

Checks if the package is loaded onto the search path or not.

Usage


## S3 method for class 'Package'
getDate(this, ...)

Arguments

...
Not used.

Value

TRUE if the packages has been loaded, otherwise FALSE.

Author(s)

Henrik Bengtsson

See Also

For more information see Package.

Examples


  pkg <- Package("base")
  print(isLoaded(pkg))       # TRUE

[Package R.oo version 1.26.0 Index]