getHowToCite.Package {R.oo}	R Documentation

Gets the citation of this package

Description

Gets the citation of this package. If text file HOWTOCITE exists in package root directory, then its contents is retrieved, otherwise citation for the package is retrieved.

Usage


## S3 method for class 'Package'
getHowToCite(this, newline="\n", ...)

Arguments

newline
The character string to collapse lines in the file.

...
Not used.

Value

Returns a character string.

Author(s)

Henrik Bengtsson

See Also

For more information see Package.

[Package R.oo version 1.26.0 Index]