getInstantiationTime.BasicObject {R.oo}	R Documentation

Gets the time when the object was instantiated

Description

Gets the time when the object was instantiated (created) as a POSIXt object.

Usage


## S3 method for class 'BasicObject'
getInstantiationTime(this, ...)

Arguments

...
Not used.

Details

The instantiation timestamp is set when the object is created, and only of option "R.oo::BasicObject/instantiationTime" is TRUE.

Value

Returns a POSIXt object, which extends class POSIXct.

Author(s)

Henrik Bengtsson

See Also

For more about time formats and POSIX see DateTimeClasses. For more information see BasicObject.

[Package R.oo version 1.26.0 Index]