getMessage.InternalErrorException {R.oo}	R Documentation

Gets the message of the exception

Description

Gets the message of the exception and adds a message that clarifies that the error is likely due to an internal error and not due to the user. It also gives information how to contact the maintainer or author of the suspicious package. This information is retrieved from the DESCRIPTION file of the package. To help the package developer, information about the current version of R, the current version of the package etc are also returned so the user can easily cut'n'paste that information into a bug report.

Usage


## S3 method for class 'InternalErrorException'
getMessage(this, ...)

Arguments

...
Not used.

Value

Returns a character string.

Author(s)

Henrik Bengtsson

See Also

For more information see InternalErrorException.

[Package R.oo version 1.26.0 Index]