getMethods.Class {R.oo}	R Documentation

Returns the method names of class and its super classes

Description

Returns the method names of class and its super classes as a list.

Usage


## S3 method for class 'Class'
getMethods(this, private=FALSE, deprecated=TRUE, unique=TRUE, ...)

Arguments

private
If TRUE, private methods are also returned, otherwise only public ones are returned.

deprecated
If TRUE, deprecated methods are also returned.

unique
If TRUE, only methods that are not implemented in one of the subclasses are returned for each class.

...
Not used.

Value

Returns a named list of named character strings.

Author(s)

Henrik Bengtsson

See Also

For more information see Class.

Examples


  names <- getMethods(Exception)
  print(names)

[Package R.oo version 1.26.0 Index]