Rdoc$getRdTitle {R.oo}	R Documentation

Extracts the title string of a Rd file

Description

Extracts the title string of a Rd file corresponding the the specified method of the specified class.

Usage


## Static method (use this):
## Rdoc$getRdTitle(class, method, ...)

## Don't use the below:
## S3 method for class 'Rdoc'
getRdTitle(this, class, method, ...)

Arguments

method
The method to be search for.

class
The class the method belongs to.

...
Not used.

Value

Returns a character string.

Author(s)

Henrik Bengtsson

See Also

For more information see Rdoc.

[Package R.oo version 1.26.0 Index]