hasField.BasicObject {R.oo}	R Documentation

Checks if a field exists or not

Description

Checks if a field exists or not.

Usage


## S3 method for class 'BasicObject'
hasField(this, field, ...)

Arguments

field
vector of fields to be checked if they exists or not.

...
Not used.

Value

Returns a logical vector indicating for each field if it exists or not.

Author(s)

Henrik Bengtsson

See Also

To get the fields of an Object, see *getFields(). For more information see BasicObject.

[Package R.oo version 1.26.0 Index]