hasField.Object {R.oo}	R Documentation

Checks if a field exists or not

Description

Checks if a field exists or not.

Usage


## S3 method for class 'Object'
hasField(this, field, ...)

Arguments

field
vector of fields to be checked if they exists or not.

...
Not used.

Value

Returns a logical vector indicating for each field if it exists or not.

Author(s)

Henrik Bengtsson

See Also

To get the fields of an Object, see *getFields(). For more extensive information about the fields in an Object see *ll(). For more information see Object.

Examples


  obj <- Object()
  obj$x <- 1:100
  obj$y <- 100:1
  hasField(obj, c("x", "a", "b", "y"))

  ## Not run: 
  gives:

  [1] TRUE FALSE FALSE TRUE
  
## End(Not run)

[Package R.oo version 1.26.0 Index]