hashCode {R.oo}	R Documentation

Gets an integer hashcoded for R objects

Description

Gets an integer hashcoded for R objects.

Usage


## Default S3 method:
hashCode(object, ...)

Arguments

object
A vector or list of R objects.

...
Not used.

Details

A character string is converted into a hashcode following Java conventions by s[1]*31^(n-1) + s[2]*31^(n-2) + ... + s[n] using integer arithmetic, where s[i] is the i:th character of the string, n is the length of the string. The hash value of the empty string is zero.

For all other types, as.integer() is called by default.

Value

Returns a vector of integer's.

Author(s)

Henrik Bengtsson

[Package R.oo version 1.26.0 Index]