isOlderThan.Package {R.oo}	R Documentation

Checks if the package is older than a given version

Description

Checks if the package is older than a given version.

Usage

## S3 method for class 'Package'
isOlderThan(this, version, ...)

Arguments

version
A character string specifying a version to compare with.

...
Not used.

Value

Returns TRUE if the version of the package is less than the specified version.

Author(s)

Henrik Bengtsson

See Also

For more information see Package. *getVersion().

[Package R.oo version 1.22.0 Index]