ll {R.oo}	R Documentation

Generates a list of informative properties of all members of an environment

Description

Generates a list of informative properties of all members of an environment.

Usage


## Default S3 method:
ll(pattern=".*", ..., private=FALSE, properties=getOption("R.oo::ll/properties",
  c("data.class", "dimension", "objectSize")), sortBy=NULL, decreasing=FALSE,
  envir=parent.frame())

Arguments

pattern
Regular expression pattern specifying which members to return. If ".*", all names are matched.

...
A named vector of format functionName=value, where functionName() will be called on each member found. If the result matches the value, the member is returned, otherwise not.

private
If TRUE, also private members, i.e. members with a name starting with a . (period), will be listed, otherwise not.

properties
Names of properties to be returned. There must exist a function with the same name, because it will be called. This way one can extract any type of property by defining new methods.

sortBy
Name or index of column (property) to be sorted by. If NULL, the objects are listed in the order they are found.

decreasing
A logical indicating whether the sorting should be done in increasing or decreasing order.

envir
An environment, a search path index or a name of a package to be scanned.

Value

Returns a data.frame containing information about all the members.

Default properties returned

It is possible to set the default value of argument properties by setting option "R.oo::ll/properties", e.g. options("R.oo::ll/properties"=c("data.class", "dimension")). If this option is not set when the package is loaded, it is set to c("data.class", "dimension", "objectSize").

Author(s)

Henrik Bengtsson

See Also

ls.str and ll.Object().

Examples


  ## Not run: 
   To list all objects in .GlobalEnv:
   > ll()
   		     member data.class dimension objectSize
   1                *tmp*     Person         1         428
   2  as.character.Person   function      NULL        1208
   3              country  character         1          44
   4        equals.Person   function      NULL        2324
   5             filename  character         1          84
   6               getAge   function      NULL         372
   7        getAge.Person   function      NULL         612
   8       getName.Person   function      NULL         628
   9      hashCode.Person   function      NULL        1196
   10        last.warning       list         1         192
   11                 obj     Person         1         428
   12              Person      Class      NULL        2292
   13              setAge   function      NULL         372
   14       setAge.Person   function      NULL        2088
   15             setName   function      NULL         372
   16      setName.Person   function      NULL         760
   17   staticCode.Person   function      NULL        2372

   To list all functions in the methods package:
   ll(mode="function", envir="methods")

   To list all numeric and character object in the base package:
   ll(mode=c("numeric", "character"), envir="base")

   To list all objects in the base package greater than 40kb:
   subset(ll(envir="base"), objectSize > 40000)
  
## End(Not run)

[Package R.oo version 1.26.0 Index]