ll.Object {R.oo}	R Documentation

Generates a list of informative properties of all members of an Object

Description

Generates a list of informative properties of all members of an Object.

Usage


## S3 method for class 'Object'
ll(this, ...)

Arguments

...
Any arguments accepted by the underlying function ll().

Value

Returns a data.frame containing information about all the members.

Author(s)

Henrik Bengtsson

See Also

ll.default(). For more information see Object.

Examples


  obj <- Object()
  obj$x <- 1:100
  obj$y <- 100:1
  ll(obj)

  ## Not run: 
  gives:

    member data.class dimension objectSize
  1      x    numeric       100        424
  2      y    numeric       100        424
  
## End(Not run)

[Package R.oo version 1.26.0 Index]