ll.Package {R.oo}	R Documentation

Generates a list of informative properties of all members of the package

Description

Generates a list of informative properties of all members of the package.

If the package is not loaded, it will be loaded, the members will be retrieved and then the package will be unloaded again.

Usage


## S3 method for class 'Package'
ll(this, envir=pos.to.env(getPosition(this)), ...)

Arguments

...
Not used.

Details

Note the difference from the default definition of ll() as inherited from the Object class. Here ll() has been redefined to list the members of the package, i.e. the members in the environment on search path that belongs to the package, whereas the original usage was to list the members of the Object. However, since the internal members of Package object is not of interest we decided on this definition instead.

Value

Returns a data.frame.

Author(s)

Henrik Bengtsson

See Also

For more information see Package.

Examples


  ## Not run: ll(R.oo)

[Package R.oo version 1.26.0 Index]