Object$load {R.oo}	R Documentation

Static method to load an Object from a file or a connection

Description

Static method to load an Object from a file or a connection, which previously have been saved using save() of class Object.

Usage


## Static method (use this):
## Object$load(file, path=NULL, ...)

## Don't use the below:
## S3 method for class 'Object'
load(static, file, path=NULL, ...)

Arguments

file
Filename or connection from where to read the Object.

path
The path where the file exists.

...
Not used.

Details

Please note that no constructors are called when an Object is loaded and neither is any static class code called.

Value

Returns a reference to the loaded Object.

Type control

Typically this static method is called as <Object>$load(...) where <Object> is any Object class. When an Object has been loaded, it is verified that it inherits from <Object>. If it does not, an exception is thrown. Thus, Object$load(...) will load any Object, but MyClass$load(...) will only load an Object that inherits from MyClass. If loaded object is not of correct class, an exception is thrown.

Troubleshooting

Due to a bug, likely in R itself, one can not specify the file argument by its name, i.e. Object$load(file="foo.RData") will not work, but Object$load("foo.RData") work just fine.

Author(s)

Henrik Bengtsson

See Also

*save() and save(), load(). For more information see Object.

Examples

## Not run: For a complete example see help(Object).

[Package R.oo version 1.26.0 Index]