newInstance.BasicObject {R.oo}	R Documentation

Creates a new instance of the same class as this object

Description

Creates a new instance of the same class as this object.

Usage


## S3 method for class 'BasicObject'
newInstance(this, ...)

Arguments

...
Arguments passed to the constructor of the corresponding BasicObject class.

Value

Returns a reference to an instance of BasicObject or a subclass thereof.

Author(s)

Henrik Bengtsson

See Also

newInstance.Object(). newInstance.Class(). For more information see BasicObject.

[Package R.oo version 1.26.0 Index]