objectSize.Object {R.oo}	R Documentation

Gets the size of the Object in bytes

Description

Gets the size of the Object in bytes by summing the sizes of all its members. For this reason, the size of memory the Object actually allocates might vary slightly.

Usage


## S3 method for class 'Object'
objectSize(this, ...)

Arguments

...
Not used.

Value

Returns an integer specifying the size of the object in number of bytes.

Author(s)

Henrik Bengtsson

See Also

To clear fields that are declared cached, see *clearCache(). object.size. For more information see Object.

Examples


  obj <- Object()
  obj$x <- 1:100
  obj$y <- 100:1
  objectSize(obj)   # 856

[Package R.oo version 1.26.0 Index]