print.Object {R.oo}	R Documentation

Prints an Object

Description

For all objects of class Object, this method will print the value of as.character() of the object. Note that this function is not called if the argument is not an object of class Object.

Usage


## S3 method for class 'Object'
print(x, ...)

Author(s)

Henrik Bengtsson

See Also

print.default() For more information see Object.

Examples

## Not run: For a complete example see help(Object).

[Package R.oo version 1.26.0 Index]