save.Object {R.oo}	R Documentation

Saves an Object to a file or a connection

Description

Saves an Object to a file or a connection.

Usage


## S3 method for class 'Object'
save(this, file=NULL, path=NULL, compress=TRUE, ..., safe=TRUE)

Arguments

file
Filename or connection to which the Object should be saved. If NULL, the Object will be save to a file named "{class name}.{memory location}.RData", e.g. "Object.26979608.RData".

path
The path where the file should be saved.

compress
If TRUE, the file is compressed to, otherwise not.

...
Other arguments accepted by save() in the base package.

safe
If TRUE and file is a file, then, in order to lower the risk for incomplete files, the object is first written to a temporary file, which is then renamed to the final name.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

*load() and save(), load(). For more information see Object.

Examples

## Not run: For a complete example see help(Object).

[Package R.oo version 1.26.0 Index]