Rdoc$setManPath {R.oo}	R Documentation

Sets the path to the directory where the Rd files should be saved

Description

Sets the path to the directory where the Rd files should be saved. By default the path is ../man/ assuming that the current directory is ../R/, which is where source files commonly are placed.

Usage


## Static method (use this):
## Rdoc$setManPath(path="../man/", ...)

## Don't use the below:
## S3 method for class 'Rdoc'
setManPath(this, path="../man/", ...)

Arguments

...
Not used.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

*getManPath() For more information see Rdoc.

[Package R.oo version 1.26.0 Index]