staticCode.Object {R.oo}	R Documentation

Method that will be call each time a new instance of a class is created

Description

Method that will be call each time a new instance of a class is created.

By extending this method it is possible to have static code that is called each time a new instance of a class is created.

Usage


## S3 method for class 'Object'
staticCode(static, ...)

Arguments

static
The static instance of this class.

...
Not used.

Details

The method extend() in the Object class will call this method just before returning and it will pass the static instance of the class as a reference. Note that method should never be called explicitly.

Any value returned by this method will be ignored.

Value

Returns nothing.

Author(s)

Henrik Bengtsson

See Also

For more information see Object.

Examples

## Not run: For a complete example see help(Object).

[Package R.oo version 1.26.0 Index]