trim {R.oo}	R Documentation

Converts to a string and removes leading and trailing whitespace

Description

Converts to a string and removes leading and trailing whitespace.

Usage

## Default S3 method:
trim(object, ...)

Arguments

object
A vector of R objects to be trimmed.

...
Not used.

Value

Returns a vector of character strings.

Author(s)

Henrik Bengtsson

[Package R.oo version 1.22.0 Index]