update.Package {R.oo}	R Documentation

Deprecated: Updates the package is a newer version is available

Description

This method is defunct. Use update.packages instead.

Usage

## S3 method for class 'Package'
update(object, contribUrl=c(getContribUrl(object), getDevelUrl(object)), force=FALSE,
  reload=TRUE, verbose=TRUE, ...)

Arguments

contribUrl
The URL from where the package can be installed and updated. By default the URL according to the DESCRIPTION is assumed. If the URL is missing, CRAN is assumed.

force
If TRUE, the package will reinstalled even if it is up to date according to the version number.

verbose
If TRUE, more detailed information is returned.

...
Not used.

Value

Returns (invisibly) TRUE if the package was updated, otherwise FALSE.

Author(s)

Henrik Bengtsson

See Also

update.packages. For more information see Package.

[Package R.oo version 1.20.0 Index]