dimension {R.oo}	R Documentation

Gets the dimension of the object

Description

Gets the dimension of the object similar to what dim() does, but instead of NULL it will return the length of a vector. If a function is passed, NULL is returned.

Usage

## Default S3 method:
dimension(object, ...)

Arguments

object
The object for which the dimension should be obtained.

...
Not used.

Value

Returns an integer vector or NULL.

Author(s)

Henrik Bengtsson

See Also

ll.default(). dim() and length().

Examples

  dimension(matrix(1:100, ncol=10))     # 10 10
  dimension(1:14)                       # 14
  dimension(data.frame(a=1:10, b=10:1)) # 10  2
  dimension(print)                      # NULL

[Package R.oo version 1.22.0 Index]